refactor: some corrections
