+++ Example/PrebidDemo/PrebidDemoTests/Constants.swift
@@ -16,6 +16,7 @@
 struct Constants {
 
     // MARK: - AppNexus
+    static let PBS_APPNEXUS_HOST_URL = "https://prebid.adnxs.com/pbs/v1/openrtb2/auction"
     static let PBS_ACCOUNT_ID_APPNEXUS = "bfa84af2-bd16-4d35-96ad-31c6bb888df0"
     static let PBS_CONFIG_ID_300x250_APPNEXUS = "6ace8c7d-88c0-4623-8117-75bc3f0a2e45"
     static let PBS_CONFIG_ID_INTERSTITIAL_APPNEXUS = "625c6125-f19e-4d5b-95c5-55501526b2a4"
@@ -29,7 +30,7 @@ struct Constants {
     static let PBS_INVALID_ACCOUNT_ID_APPNEXUS = "bfa84af2-bd16-4d35-96ad-ffffffffffff"
     static let PBS_INVALID_CONFIG_ID_APPNEXUS = "6ace8c7d-88c0-4623-8117-ffffffffffff"
     
-    // MARK: - AppNexus
+    // MARK: - Rubicon
     static let PBS_RUBICON_ACCOUNT_ID = "1001"
     static let PBS_CONFIG_ID_300x250_RUBICON = "1001-1"
     static let PBS_CONFIG_ID_INTERSTITIAL_RUBICON = "1001-1"
