+++ Example/PrebidDemo/PrebidDemoTests/StubbingHandler.swift
@@ -1,0 +1,47 @@
+/*   Copyright 2018-2021 Prebid.org, Inc.
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ */
+
+import Foundation
+import TestUtils
+
+public class StubbingHandler {
+    
+    private init() {}
+    
+    public static let shared = StubbingHandler()
+    
+    public func turnOn() {
+        PBHTTPStubbingManager.shared().enable()
+        PBHTTPStubbingManager.shared().ignoreUnstubbedRequests = true
+        PBHTTPStubbingManager.shared().broadcastRequests = true
+    }
+    
+    public func turnOff() {
+        PBHTTPStubbingManager.shared().disable()
+        PBHTTPStubbingManager.shared().removeAllStubs()
+        PBHTTPStubbingManager.shared().broadcastRequests = false
+        PBHTTPStubbingManager.shared().ignoreUnstubbedRequests = true
+    }
+    
+    public func stubRequest(with responseName: String, requestURL: String) {
+        let currentBundle = Bundle(for: TestUtils.PBHTTPStubbingManager.self)
+        let baseResponse = try? String(contentsOfFile: currentBundle.path(forResource: responseName, ofType: "json") ?? "", encoding: .utf8)
+        let requestStub = PBURLConnectionStub()
+        requestStub.requestURL = requestURL
+        requestStub.responseCode = 200
+        requestStub.responseBody = baseResponse
+        PBHTTPStubbingManager.shared().add(requestStub)
+    }
+}
